+++ src/Recyclarr/Config/EmptyYamlException.cs
@@ -1,0 +1,5 @@
+namespace Recyclarr.Config;
+
+public class EmptyYamlException : Exception
+{
+}