+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Do not warn about empty configuration YAML files when they aren't really empty.
+
 ## [3.1.0] - 2022-12-10
 
 ### Changed